refactor: 의존성 주입이 자동으로 수행되는 Activity 클래스를 생성한다.
